requirements.txt
streamlit==1.32.0
google-generativeai
jinja2==3.1.3
python-dotenv==1.0.1
